+++ gym/monitoring/__init__.py
@@ -1,9 +1,2 @@
-from gym.monitoring.monitor_manager import (
-    _open_monitors,
-    detect_training_manifests,
-    load_env_info_from_manifests,
-    load_results,
-    MonitorManager,
-)
 from gym.monitoring.stats_recorder import StatsRecorder
 from gym.monitoring.video_recorder import VideoRecorder
